Abort on yt-dlp DownloadError
Some yt-dlp extractors die with this exception when the url doesn't actually have any video on it.

For instance, NYT articles sometimes have videos that yt-dlp will look for, and then crash if not found.

In this situation, we should continue with preprocessing under the assumption that we are summarizing text, not video.